requirements.txt
mysql-connector-python
requests 
bs4 
sqlalchemy